added styles to detail view
PASÉ EL COMPONENTE DETAIL A VIEWS!!!
