+++ Client/src/views/Detail/Detail.module.css
@@ -1,0 +1,91 @@
+.backButton {
+  position: absolute;
+  top: 20px;
+  left: 20px;
+  border-radius: 50px;
+
+  width: 80px;
+  height: 80px;
+  padding: 0;
+  font-size: 40px;
+  line-height: 0;
+  letter-spacing: 0;
+}
+
+.detailContainer {
+  width: 800px;
+
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  align-items: flex-start;
+
+  margin: auto;
+}
+
+.imgContainer {
+  width: 350px;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+}
+
+/* Estilos para la imagen principal (grande) */
+.mainImage {
+  width: 350px;
+  height: 500px;
+
+  object-fit: cover;
+  object-position: center;
+
+  border: 1px solid #ccc;
+  border-radius: 2px;
+}
+
+/* Estilos para las imágenes en miniatura (thumbnails) */
+.thumbnail {
+  width: 80px;
+  height: 80px;
+  object-fit: cover;
+  object-position: center;
+  margin: 5px 5px 0 0;
+  border: 1px solid #ccc;
+  border-radius: 2px;
+}
+
+.detailInfo {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+
+  text-align: left;
+  font-size: 16px;
+
+  width: 350px;
+}
+.detailName {
+  font-size: 26px;
+  font-weight: 700;
+  margin: 0 0 30px 0;
+}
+.detailPrice {
+  font-size: 26px;
+  font-weight: 700;
+}
+
+.detailButtons button {
+  width: 100%;
+  border-radius: 2px;
+  font-weight: 500;
+}
+.favButton {
+  background-color: #fff;
+  border: solid 1px #333;
+  color: #333;
+}
+.cartButton {
+  margin-top: 10px;
+  background-color: #000;
+  border: solid 1px #333;
+  color: #fff;
+}
